requirements.txt
numpy==1.22.4
opencv_python==4.7.0.72
Pillow==9.1.1
Pillow==9.5.0
streamlit==1.9.2
